style(gap-234): fuel-core clippy build-breaker lints, per-site by hand (no --fix)
GAP-234 fuel-core leg — the 3 MachineApplicable clippy lints whose mechanical --fix
breaks the build, fixed BY HAND per-site (the compiler is the arity gate):

  chunks_exact_to_as_chunks (92 sites): `.chunks_exact(N)` -> `.as_chunks::<N>().0.iter()`.
    CORRECTION to the row's "fuel-core 2-part": measured across ALL 92, only 3 are
    2-part; 89 are 1-part. The 89 index `[c[0], c[1], ...]` / `.map(|c| [c[i]])`, which
    compiles on `&[u8;N]` unchanged (all lazy_quantized_* loaders, lazy.rs decoders, the
    vision models). The 3 two-part (lazy.rs x2, lazy_bert.rs) add
    `chunk.try_into().unwrap()/.expect()` -> `*chunk` (a fallible conversion that can
    never fail, removed). My earlier "uniformly 2-part" generalized from one site.
  int_plus_one (2, lazy_xlm_roberta): `seq + cfg.pad_token_id as usize + 1 <= max` ->
    `seq + (cfg.pad_token_id as usize) < max` -- parenthesize the cast so `<` is not read
    as generic args (guard the parse, not the intent).
  unnecessary_cast (4, lazy.rs/lazy_convmixer/lazy_z_image x2): remove redundant
    same-type casts, value-preserving.

NOT an allow -- the lints are right, only clippy's auto-fix is wrong in context. Every
site was classified by the compiler: a 1-part misclassification fails E0277, so a clean
build proves the split. fuel-core now shows 0/3 of these lints; the remaining 11
chunks_exact are all fuel-quantized's (LAST commit, compiler-only gate).

Verified: cargo test -p fuel-core -> lib 1382 passed; 0 failed; 12 ignored;
1474 passed / 0 failed / 17 ignored total.

+++ fuel-core/src/lazy_convmixer.rs
@@ -375,7 +375,7 @@ mod tests {
             mean[i] += nb() * 0.05;
             var[i] += nb().abs() * 0.05;
         }
-        BatchNormParams::from_raw(&gain, &bias, &mean, &var, cfg.bn_eps as f64)
+        BatchNormParams::from_raw(&gain, &bias, &mean, &var, cfg.bn_eps)
     }
 
     fn tiny_weights(cfg: &ConvMixerConfig) -> ConvMixerWeights {
